Some more

+++ src/euler/euler002.rs
@@ -2,7 +2,7 @@ const MAX: u64 = 4_000_000;
 
 
 
-pub fn main() -> u64 {
+pub fn main() -> i64 {
     let mut a: u64 = 1;
     let mut b: u64 = 2;
 
@@ -12,7 +12,7 @@ pub fn main() -> u64 {
         let c = a + b;
 
         if c >= MAX {
-            return sum
+            return sum as i64
         }
         if c % 2 == 0 {
             sum += c;
