+++ src/euler/euler012.rs
@@ -1,5 +1,5 @@
 
 use utils::{triangles, divisors};
-pub fn main() -> u64 {
-    triangles().filter(|&n| divisors(n).count() > 500).next().unwrap()
+pub fn main() -> i64 {
+    triangles().filter(|&n| divisors(n).count() > 500).next().unwrap() as i64
 }
